+++ src/create-route-middleware.js
@@ -1,6 +1,6 @@
 import { curry } from 'ramda';
-import getIsEnabled from './utils/get-is-enabled';
-import updateFeatures from './utils/updateFeaturesWithParams';
+import getIsEnabled from './get-is-enabled';
+import updateFeatures from './update-features-with-query';
 import { parse } from 'url';
 
 const setStatus = (res, isEnabled) =>
