+++ src/index.js
@@ -1,13 +1,15 @@
 import withFeatures from './with-features';
 import configureFeature from './configure-feature';
 import createRouteMiddleware from './create-route-middleware';
-import getIsEnabled from './utils/get-is-enabled';
-import getEnabled from './utils/get-enabled';
+import getIsEnabled from './get-is-enabled';
+import getEnabled from './get-enabled';
+import updateFeaturesWithQuery from './update-features-with-query';
 
 export {
   withFeatures,
   configureFeature,
   createRouteMiddleware,
   getIsEnabled,
-  getEnabled
+  getEnabled,
+  updateFeaturesWithQuery
 };
