--- src/utils/index.js
@@ -1,7 +1,0 @@
-import getIsEnabled from './get-is-enabled';
-import getEnabled from './get-enabled';
-
-export default {
-  getIsEnabled,
-  getEnabled
-};
