--- src/utils/test.js
@@ -1,3 +1,0 @@
-import './get-is-enabled/test';
-import './get-enabled/test';
-import './updateFeaturesWithParams/test';
